refactor: 플레이어에 wantToDraw 제거 요청마다 처리하도록 변경
